Switch transfer values from size_t to uint64_t

+++ src/Transfers/TransfersSynchronizer.cpp
@@ -155,7 +155,7 @@ void TransfersSyncronizer::save(std::ostream& os) {
   CryptoNote::BinaryOutputStreamSerializer s(stream);
   s(const_cast<uint32_t&>(TRANSFERS_STORAGE_ARCHIVE_VERSION), "version");
 
-  size_t subscriptionCount = m_consumers.size();
+  uint64_t subscriptionCount = m_consumers.size();
 
   s.beginArray(subscriptionCount, "consumers");
 
@@ -172,7 +172,7 @@ void TransfersSyncronizer::save(std::ostream& os) {
 
     std::vector<AccountPublicAddress> subscriptions;
     consumer.second->getSubscriptions(subscriptions);
-    size_t subCount = subscriptions.size();
+    uint64_t subCount = subscriptions.size();
 
     s.beginArray(subCount, "subscriptions");
 
@@ -235,7 +235,7 @@ void TransfersSyncronizer::load(std::istream& is) {
   std::vector<ConsumerState> updatedStates;
 
   try {
-    size_t subscriptionCount = 0;
+    uint64_t subscriptionCount = 0;
     s.beginArray(subscriptionCount, "consumers");
 
     while (subscriptionCount--) {
@@ -260,7 +260,7 @@ void TransfersSyncronizer::load(std::istream& is) {
         }
 
         // load subscriptions
-        size_t subCount = 0;
+        uint64_t subCount = 0;
         s.beginArray(subCount, "subscriptions");
 
         while (subCount--) {
